backend_ai/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
psycopg2-binary
python-multipart
google-generativeai
pillow
requests
